requirements.txt
Flask>=1.1.0
python-dotenv>=0.14.0